+++ site/guidelines/overview/index.jsx
@@ -15,7 +15,7 @@ import PageBody from 'app_modules/site/components/page/body';
 import { prefix as pf } from 'app_modules/ui/util/component';
 
 export default (
-  <PageBody anchorTitle="Design" contentClassName={pf('site-design site-design-layout')}>
+  <PageBody anchorTitle="Guidelines" contentClassName={pf('site-guidelines site-guidelines-layout')}>
     <div className={pf('container--large')}>
       <p className="site-text-introduction">The Lightning Design System reflects the patterns and components that underpin the Salesforce product. These patterns and components provide a unified language and consistent look and feel when designing apps and products within the Salesforce ecosystem.</p>
       <h2 className="site-text-heading--large">Design Principles</h2>
@@ -23,7 +23,7 @@ export default (
       <ul className={pf('grid wrap grid--pull-padded-large text-align--center p-vertical--x-large')}>
         <li className={pf('col--padded-large size--1-of-1 small-size--1-of-2')}>
           <div className="grid-card">
-            <img src="/assets/images/design/img-clarity.svg" alt="" />
+            <img src="/assets/images/guidelines/img-clarity.svg" alt="" />
             <dl>
               <dt className="site-text-heading--label-weak-large">Clarity</dt>
               <dd>
@@ -35,7 +35,7 @@ export default (
         </li>
         <li className={pf('col--padded-large size--1-of-1 small-size--1-of-2')}>
           <div className="grid-card">
-            <img src="/assets/images/design/img-efficiency.svg" alt="" />
+            <img src="/assets/images/guidelines/img-efficiency.svg" alt="" />
             <dl>
               <dt className="site-text-heading--label-weak-large">Efficiency</dt>
               <dd>
@@ -47,7 +47,7 @@ export default (
         </li>
         <li className={pf('col--padded-large size--1-of-1 small-size--1-of-2')}>
           <div className="grid-card">
-            <img src="/assets/images/design/img-consistency.svg" alt="" />
+            <img src="/assets/images/guidelines/img-consistency.svg" alt="" />
             <dl>
               <dt className="site-text-heading--label-weak-large">Consistency</dt>
               <dd>
@@ -59,7 +59,7 @@ export default (
         </li>
         <li className={pf('col--padded-large size--1-of-1 small-size--1-of-2')}>
           <div className="grid-card">
-            <img src="/assets/images/design/img-beauty.svg" alt="" />
+            <img src="/assets/images/guidelines/img-beauty.svg" alt="" />
             <dl>
               <dt className="site-text-heading--label-weak-large">Beauty</dt>
               <dd>