+++ RELEASENOTES.md
@@ -152,7 +152,7 @@ The following features are being removed in this release:
 
 ## Release [1.0.5] - May 26, 2016
 
-- Adding a background color to icons from the [custom set](/resources/icons/#custom) is now done using the `slds-icon-custom-customXX` class (e.g. `slds-icon-custom-custom25`), instead of `slds-icon-custom-XX` (deprecated). This change addresses [#203](https://github.com/salesforce-ux/design-system/issues/203)
+- Adding a background color to icons from the [custom set](/icons/#custom) is now done using the `slds-icon-custom-customXX` class (e.g. `slds-icon-custom-custom25`), instead of `slds-icon-custom-XX` (deprecated). This change addresses [#203](https://github.com/salesforce-ux/design-system/issues/203)
 - Improved accessibility of the [breadcrumbs component](/components/breadcrumbs/)
 - The Design System now allows developers to load Salesforce webfonts using their preferred method instead of the default `@font-face` rules. To disable the default font loading method, set `$slds-load-webfonts: false;` before importing the Design System Sass file.
 - The project is now compatible with Node.js 6
@@ -207,7 +207,7 @@ The following site enhancements are now available:
 - Improved Component Overview tables for more narrow readability
 - Improved status labels. Now associated with colors and contain more information
 - Markup and Style guidelines with BEM information - [Markup and Style Guidelines](/getting-started/markup-and-style/)
-- Resources Information - [Resource Links](/resources/links/)
+- Resources Information - [Resource Links](/articles/)
 - Updated documentation and accessibility information for tabs - [Tabs](/components/tabs/)
 - Improved Grid System documentation - [Grid System](/components/utilities/grid/)
 - Copy updates on components to represent Voice & Tone Guidelines